+++ ast/src/ast.rs
@@ -1,6 +1,6 @@
 use crate::Term;
 use indexmap::IndexMap;
-use num_rational::{BigRational, Rational64};
+use num_rational::Rational64;
 
 #[derive(Debug, Clone, PartialEq, Hash)]
 pub struct FunDef {
@@ -47,7 +47,7 @@ pub enum Op {
         m: Rational64,
     },
     TransposeM {
-        m: BigRational,
+        m: Rational64,
     },
     TransposeA {
         a: Rational64,
